Break daemon ↔ plotting cycle: move harvester config helpers to chia.util
- Add chia.util.harvester_config with add_plot_directory, get_plot_directories,
  remove_plot_directory (extracted from chia.plotting.util)
- Update daemon, harvester, plotters, cmds, simulator, and tests to import
  from chia.util.harvester_config
- Remove daemon's dependency on plotting in tach.toml (cycle broken)
