+++ chia/harvester/harvester.py
@@ -28,11 +28,8 @@ from chia.plotting.util import (
     PlotRefreshEvents,
     PlotRefreshResult,
     PlotsRefreshParameter,
-    add_plot_directory,
     get_harvester_config,
-    get_plot_directories,
     remove_plot,
-    remove_plot_directory,
     update_harvester_config,
 )
 from chia.protocols.outbound_message import NodeType
@@ -40,6 +37,7 @@ from chia.rpc.rpc_server import StateChangedProtocol, default_get_connections
 from chia.server.server import ChiaServer
 from chia.server.ws_connection import WSChiaConnection
 from chia.util.cpu import available_logical_cores
+from chia.util.harvester_config import add_plot_directory, get_plot_directories, remove_plot_directory
 
 log = logging.getLogger(__name__)
